Add files via upload
Prototyping with paper as "pins".

This program will:

Capture a frame of the blank surface. Prompt user to turn on the map and place the papers. Capture a frame of the surface with the papers after the map projection is turned off. Create an alpha mask of the papers. Save this as an image with a transparent background (papers_alpha_mask.png). Now we can use this transparent background image to overlay onto the map during projection. Note that the placement of the papers on the map needs to be precise to match the original position where people placed them.