Add files via upload
Difference Calculation: Calculate the difference between the frames and threshold this to create the binary mask.

Apply Mask: Apply the binary mask to one of the original color frames or the color difference image to retain the color information in the areas of interest.

generate_unique_filename Function: This function generates a unique filename by appending numbers to the base name if the filename already exists in the specified directory. It ensures that no existing files are overwritten.

File Saving: In the process_and_save_frames function, the unique filenames for the difference image and alpha mask are generated before saving them.